Enhance FieldBuilder with Promise Support for Relations
Refactor the FieldBuilder to support asynchronous resolution of relation targets,
enabling circular reference resolution without compile-time dependencies.

Changes:
- Add SyncOrAsync<T> helper type for Promise support
- Update RelationAdder to work with async target functions
- Clean up comments and simplify implementation
- Update CircularNodeExample to demonstrate the functionality
